+++ js/main.js
@@ -3,7 +3,7 @@ import './miniature.js';
 import './picture-modal.js';
 import './upload-modal.js';
 import './validate-form.js';
-import './slider.js';
+import './resize-img.js';
 import {createPhotoData} from './create-foto-data.js';
 import {renderThumbnails} from './miniature.js';
 import {thumbnailClickHendler} from './picture-modal.js';
